+++ src/components/footer/components/NavBar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const Pages = () => {
+const NavBar = () => {
   return (
     <div>
       <ul>
@@ -12,4 +12,4 @@ const Pages = () => {
   );
 };
 
-export default Pages;
+export default NavBar;
